renaming files, updated profile controller to use exports and profileRoutes file and apdated authController
