+++ models/DatabaseModel.js
@@ -2,7 +2,7 @@ const mongoose = require('mongoose');
 const { User } = require('./userModel');
 const { Profile } = require('./profileModel');
 const { Drug } = require('./drugModel');
-// const { Medication } = require('../models/MedicationModel');
+const { Medication } = require('./medicationModel');
 require('dotenv').config();
 
 const db = mongoose.createConnection(process.env.MONGODB_URL);
@@ -11,7 +11,7 @@ const models = {
     User: db.model('User', User.schema, 'users'),
     Drug: db.model('Drug', Drug.schema, 'drugs'),
     Profile: db.model('Profile', Profile.schema, 'profiles'),
-    // Medication: db.model('Medication', Medication.schema, 'medications'),
+    Medication: db.model('Medication', Medication.schema, 'medications'),
 };
 
 module.exports = models;
