+++ models/abuddyModel.js
@@ -1,7 +1,5 @@
 const mongoose = require('mongoose');
 
-
-
 const aBuddySchema = new mongoose.Schema({
     
     aBuddyFirstName: {
