+++ routes/DatabaseRoutes.js
@@ -1,6 +1,5 @@
-const express = require('express');
-const router = express.Router();
-const DatabaseController = require('../controllers/DatabaseController');
+const router = require('express').Router();
+const DatabaseController = require('../controllers/databaseController');
 
 router.get('/getDatabases', (req, res, next) => {
 	DatabaseController.getDatabases(req, res, next);
