+++ routes/DrugRoutes.js
@@ -1,5 +1,4 @@
-const express = require('express');
-const router = express.Router();
+const router = require('express').Router();
 const DrugController = require('../controllers/drugController');
 
 router.get('/', (req, res, next) => {
