+++ routes/UserRoutes.js
@@ -1,5 +1,4 @@
-const express = require('express');
-const router = express.Router();
+const router = require('express').Router();
 const UserController = require('./UserController');
 const { authenticate, requireRole } = require('../middlewares/auth');
