+++ routes/profileRoutes.js
@@ -1,5 +1,4 @@
-const express = require('express');
-const router = express.Router();
+const router = require('express').Router();
 const ProfileController = require('../controllers/profileController');
 
 router.get('/', (req, res, next) => {
